refactor(profile): Move shared Geo interfaces to 'common'
Changes:
- Created 'common' package
- Moved:
  * IdableNameableGeo
  * IdentifiableGeo
  * NameableGeo
