+++ mp-profile/src/main/java/ru/pavbatol/myplace/geo/common/IdableNameableGeo.java
@@ -1,4 +1,4 @@
-package ru.pavbatol.myplace.geo;
+package ru.pavbatol.myplace.geo.common;
 
 public interface IdableNameableGeo extends IdentifiableGeo, NameableGeo {
 }
