+++ mp-profile/src/main/java/ru/pavbatol/myplace/geo/common/NameableGeo.java
@@ -1,4 +1,4 @@
-package ru.pavbatol.myplace.geo;
+package ru.pavbatol.myplace.geo.common;
 
 public interface NameableGeo {
     String getName();
