+++ mp-profile/src/main/java/ru/pavbatol/myplace/geo/district/dto/DistrictDto.java
@@ -2,7 +2,7 @@ package ru.pavbatol.myplace.geo.district.dto;
 
 import lombok.Value;
 import ru.pavbatol.myplace.app.util.Marker;
-import ru.pavbatol.myplace.geo.IdentifiableGeo;
+import ru.pavbatol.myplace.geo.common.IdentifiableGeo;
 import ru.pavbatol.myplace.geo.region.dto.RegionDto;
 
 import javax.validation.constraints.NotNull;
