+++ mp-profile/src/main/java/ru/pavbatol/myplace/geo/management/service/DataImportServiceImpl.java
@@ -6,9 +6,9 @@ import org.springframework.lang.Nullable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
-import ru.pavbatol.myplace.geo.IdableNameableGeo;
-import ru.pavbatol.myplace.geo.IdentifiableGeo;
-import ru.pavbatol.myplace.geo.NameableGeo;
+import ru.pavbatol.myplace.geo.common.IdableNameableGeo;
+import ru.pavbatol.myplace.geo.common.IdentifiableGeo;
+import ru.pavbatol.myplace.geo.common.NameableGeo;
 import ru.pavbatol.myplace.geo.city.model.City;
 import ru.pavbatol.myplace.geo.city.repository.CityRepository;
 import ru.pavbatol.myplace.geo.country.model.Country;
