+++ mp-profile/src/main/java/ru/pavbatol/myplace/geo/region/dto/RegionDto.java
@@ -2,7 +2,7 @@ package ru.pavbatol.myplace.geo.region.dto;
 
 import lombok.Value;
 import ru.pavbatol.myplace.app.util.Marker;
-import ru.pavbatol.myplace.geo.IdentifiableGeo;
+import ru.pavbatol.myplace.geo.common.IdentifiableGeo;
 import ru.pavbatol.myplace.geo.country.dto.CountryDto;
 
 import javax.validation.constraints.NotNull;
