+++ mp-profile/src/main/java/ru/pavbatol/myplace/geo/region/model/Region.java
@@ -3,7 +3,7 @@ package ru.pavbatol.myplace.geo.region.model;
 import lombok.*;
 import lombok.experimental.Accessors;
 import lombok.experimental.FieldDefaults;
-import ru.pavbatol.myplace.geo.IdableNameableGeo;
+import ru.pavbatol.myplace.geo.common.IdableNameableGeo;
 import ru.pavbatol.myplace.geo.country.model.Country;
 
 import javax.persistence.*;
